Move TestGenieSettingsService to Services folder
refactoring to keep code consistent
