+++ src/main/kotlin/nl/tudelft/ewi/se/ciselab/testgenie/evosuite/SettingsArguments.kt
@@ -1,6 +1,6 @@
 package nl.tudelft.ewi.se.ciselab.testgenie.evosuite
 
-import nl.tudelft.ewi.se.ciselab.testgenie.settings.TestGenieSettingsService
+import nl.tudelft.ewi.se.ciselab.testgenie.services.TestGenieSettingsService
 import nl.tudelft.ewi.se.ciselab.testgenie.settings.TestGenieSettingsState
 import nl.tudelft.ewi.se.ciselab.testgenie.toolwindow.QuickAccessParametersService
