+++ src/main/kotlin/nl/tudelft/ewi/se/ciselab/testgenie/settings/SettingsPluginConfigurable.kt
@@ -1,6 +1,7 @@
 package nl.tudelft.ewi.se.ciselab.testgenie.settings
 
 import com.intellij.openapi.options.Configurable
+import nl.tudelft.ewi.se.ciselab.testgenie.services.TestGenieSettingsService
 import javax.swing.JComponent
 
 /**
